Dropped Core Indexer in favor of IncrementalIndexer contracts
